Corregidos los Servicios de Cliente, comercial y vehiculo. Corregido javaconfig y añadido fichero de configuración para test de persistencia

+++ concesionarioRest/Servicios/src/main/java/com/atsistemas/concesionario/servicios/ServicioVehiculo.java
@@ -19,20 +19,40 @@ public class ServicioVehiculo implements VehiculoServicio {
 		this.vehiculoDao = vehiculoDao;
 	}
 
+	
 	public long alta(Vehiculo vehiculo) {
-		return ((Vehiculo) vehiculoDao.save(vehiculo)).getId();
+		// Como la implementación de save de JpaRepository devuelve un tipo
+		// <S extends T> S save(S entity), hay que hacer un cast para
+		// devolver lo que queramos, ya que recibiremos por defecto el
+		// objeto que insertemos en la tabla destino
+		if (vehiculo != null) {
+			return ((Vehiculo) vehiculoDao.save(vehiculo)).getId();
+		} else {
+			return (0L);
+		}
 	}
 
 	public long baja(long idVehiculo) {
-		return vehiculoDao.delete(idVehiculo);
+		if (idVehiculo > 0L) {
+			return vehiculoDao.delete(idVehiculo);
+		} else
+			return (0L);
 	}
 
 	public long modificacion(Vehiculo vehiculo) {
-		return ((Vehiculo) vehiculoDao.save(vehiculo)).getId();
+		// Como la implementación de save de JpaRepository devuelve un tipo
+		// <S extends T> S save(S entity), hay que hacer un cast para
+		// devolver lo que queramos, ya que recibiremos por defecto el
+		// objeto que insertemos en la tabla destino
+		if (vehiculo != null) {
+			return ((Vehiculo) vehiculoDao.save(vehiculo)).getId();
+		} else {
+			return (0l);
+		}
 	}
 
-	public Vehiculo consultaPorId(long id) {
-		return vehiculoDao.findOne(id);
+	public Vehiculo consultaPorId(long idVehiculo) {
+		return vehiculoDao.findOne(idVehiculo);
 	}
 
 	public List<Vehiculo> consultarTodos() {
